Mapper 功能完善 delete

+++ model-spring-suport/src/main/java/com/javaoffers/batis/modelhelper/fun/crud/WhereModifyFun.java
@@ -9,7 +9,7 @@ import com.javaoffers.batis.modelhelper.fun.crud.update.SmartUpdateFun;
  * @Description: where
  * @Auther: create by cmj on 2022/5/4 21:44
  */
-public interface WhereModifyFun<M,V> extends  WhereFun<M, GetterFun<M,V>,V, WhereModifyFun<M,V>> , ExecutOneFun<Long> {
+public interface WhereModifyFun<M,V> extends  WhereFun<M, GetterFun<M,V>,V, WhereModifyFun<M,V>> , ExecutOneFun<Integer> {
 
     /**
      * add patch
